add a blob script which creates data.json files
blob.js will create a data.json blob for any query.js it comes across. this minimal slice of data can be imported by a component or router component wrapper.

in the future this should just be read from the component page. the graphql server should also be part of a webpack dev server. or i need to figure out how that local graphql layer can be used to render on both dev and build. for instance, when does blob/query run in the dev/build cycle?

note this is not the way react-static likes to pair data.

why do we really need a json file? it makes the import of data a bit more explicit, rather than injecting it directly into the component file. this way you separate the injection step from your actual code.

could we do something similar with a HOC handling the fetch and data passing?