move queries to be in component file
colocation!

one issue with this is that node does not like the import statements in the component file. also, i kind of like the flexibility of having the queries in their own files.

perhaps move back to that form. can use `export` for the query so developers do not need to worry about it. what is cool is you could put a query anywhere you need, run `fetch-data` and voila, you have all of the data you need for importing in a concise format.